fix cloud-extra
git-svn-id: https://svn.java.net/svn/glassfish~svn/trunk/main@54975 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5

+++ cloud-extra/tenant-manager/src/test/java/org/glassfish/paas/tenantmanager/api/TenantManagerTest.java
@@ -133,8 +133,8 @@ public class TenantManagerTest extends ConfigApiTest {
             Assert.assertNotNull("Zero Services", tenant.getServices());
             Assert.assertNotNull("Zero Extensions", tenant.getExtensions());
             Environments environments = tenant.getEnvironments();
-            Assert.assertEquals("Next Id", 1, environments.allocateId());
-            Assert.assertEquals("Next Id", 2, environments.allocateId());
+            Assert.assertEquals("Next Id", 1L, environments.allocateId().longValue());
+            Assert.assertEquals("Next Id", 2L, environments.allocateId().longValue());
             tenantManager.setCurrentTenant("tenant2");
             tenant = tenantManager.get(Tenant.class);
             Assert.assertNotNull("currentTenant", tenant);
